# perf: Optimize insights and fingerprinting of time series