Can be reviewed commit by commit.

1. Improved `redux.core/juxt` and made our functions use it. This is a placeholder before these changes are merged and released in a redux library fork. Completely rewrote `simple-linear-regression` from kixi.stats to allocate less and use primitive math more.

2. Rewrote `fingerprinters/with-error-handling` because the current implementation produced a lot of garbage in closure allocations since `try-catch` is very sensitive to positioning and will compile into functions and closures where it can be avoided if written by hand.

Same single-table benchmark as #47424, 20k rows displayed.

```
Before:

Time per call: 1.62 s      Alloc per call: 1,983,336,344b
Time per call: 1.52 s      Alloc per call: 1,983,146,208b
Time per call: 1.20 s      Alloc per call: 1,983,140,328b
Time per call: 1.55 s      Alloc per call: 1,983,141,464b
Time per call: 1.45 s      Alloc per call: 1,983,130,136b

After:

Time per call: 507.29 ms   Alloc per call: 427,498,416b
Time per call: 615.39 ms   Alloc per call: 427,494,792b
Time per call: 615.08 ms   Alloc per call: 427,495,176b
Time per call: 503.81 ms   Alloc per call: 427,495,624b
Time per call: 523.84 ms   Alloc per call: 427,494,920b
```

<img width="270" alt="image" src="https://github.com/user-attachments/assets/395df0d7-90b4-4333-b003-4c85ad3fe7a9">